Adjust Analytics folder, switch Trends tab to All Items or Specific Item
