+++ Chronicle/Analytics/Mood/MoodAnalyticsView.swift
@@ -9,12 +9,13 @@ import SwiftUI
 import Charts
 
 struct MoodAnalyticsView: View {
+    var item: Item
     var sessions: [Session]
     @Binding var filter: DateFilter
     
     var body: some View {
         NavigationLink {
-            MoodsAnalyticsDetailsView(sessions: sessions, filter: $filter)
+            MoodsAnalyticsDetailsView(item: item, sessions: sessions, filter: $filter)
         } label: {
             VStack(alignment: .leading) {
                 HStack {
@@ -68,6 +69,7 @@ struct MoodAnalyticsView: View {
 }
 
 struct MoodsAnalyticsDetailsView: View {
+    var item: Item
     var sessions: [Session]
     @Binding var filter: DateFilter
     
@@ -80,7 +82,7 @@ struct MoodsAnalyticsDetailsView: View {
                     }
                 }
                 .pickerStyle(SegmentedPickerStyle())
-                
+                .padding(.horizontal)
                 VStack(alignment: .leading) {
                     Text("Moods")
                         .font(.title3)
@@ -91,26 +93,25 @@ struct MoodsAnalyticsDetailsView: View {
                         .fontDesign(.rounded)
                         .foregroundStyle(.secondary)
                 }
+                .padding(.horizontal)
                 VStack(alignment: .leading, spacing: 20) {
+                    ItemMoodInsightsView(item: item, sessions: sessions)
                     GroupBox("Average Weekday Mood") {
                         moodTrendChart
                             .frame(height: 200)
                     }
                     .backgroundStyle(Color(.secondarySystemGroupedBackground))
+                    .padding(.horizontal)
                     GroupBox("Moods Logged") {
                         MoodPieChartView(sessions: sessions)
                             .frame(height: 300)
                     }
                     .backgroundStyle(Color(.secondarySystemGroupedBackground))
-                    GroupBox("Top Emotions") {
-                        topEmotionsChart
-                            .frame(height: 200)
-                    }
-                    .backgroundStyle(Color(.secondarySystemGroupedBackground))
+                    .padding(.horizontal)
                 }
             }
             .animation(.default, value: sessions)
-            .padding(.horizontal)
+            
         }
         .background(Color(.systemGroupedBackground))
     }
@@ -149,18 +150,6 @@ struct MoodsAnalyticsDetailsView: View {
         }
     }
     
-    private var topEmotionsChart: some View {
-        Chart(topEmotions(), id: \.name) { emotion in
-            BarMark(
-                x: .value("Emotion", emotion.name),
-                y: .value("Count", emotion.count)
-            )
-            .cornerRadius(8.0)
-            .foregroundStyle(by: .value("Emotion", emotion.name))
-        }
-        .chartLegend(.hidden)
-    }
-    
     private func averageWeekdayMoods() -> [(weekday: String, averageMood: Double)] {
         let calendar = Calendar.autoupdatingCurrent
         let groupedSessions = Dictionary(grouping: sessionsWithMoods()) { session in
@@ -173,17 +162,6 @@ struct MoodsAnalyticsDetailsView: View {
             let averageMood = sessions.isEmpty ? 0 : totalMood / Double(sessions.count)
             return (weekday: calendar.shortWeekdaySymbols[weekday-1], averageMood: averageMood)
         }
-    }
-    
-    private func topEmotions() -> [(name: String, count: Int)] {
-        let emotions = sessions.flatMap { $0.mood?.emotions ?? [] }
-        let counts = emotions.reduce(into: [:]) { counts, emotion in
-            counts["\(emotion.emoji ?? "") \(emotion.name)", default: 0] += 1
-        }
-        return counts.sorted {
-            $0.value > $1.value
-        }.prefix(4).map { ($0.key, $0.value) }
-            .sorted { $0.name < $1.name }
     }
     
     private func sessionsWithMoods() -> [Session] {
@@ -208,7 +186,7 @@ struct MoodsAnalyticsDetailsView: View {
     
     NavigationStack {
         VStack {
-            MoodsAnalyticsDetailsView(sessions: filteredSessions, filter: $filter)
+            MoodsAnalyticsDetailsView(item: SampleData.shared.item, sessions: filteredSessions, filter: $filter)
         }
     }
     .modelContainer(SampleData.shared.container)
@@ -218,7 +196,7 @@ struct MoodsAnalyticsDetailsView: View {
     @Previewable @State var filter: DateFilter = .week
     NavigationStack {
         VStack {
-            MoodAnalyticsView(sessions: SampleData.shared.randomDatesSessions, filter: $filter)
+            MoodAnalyticsView(item: SampleData.shared.item, sessions: SampleData.shared.randomDatesSessions, filter: $filter)
         }
     }
     .modelContainer(SampleData.shared.container)
